move non-userfacing function to new file in new folder structure
